requirements.txt
pypdf
langchain
accelerate
bitsandbytes
sentence_transformers
faiss_cpu
chainlit
huggingface_hub
fastapi
jinja2
python-multipart
uvicorn
openai
tiktoken
unstructured
python-dotenv